requirements.txt
selenium~=4.10.0
driver~=0.1.0
python-dotenv~=1.0.0
psutil==5.9.5
structlog==23.2.0
aiofiles==23.2.1
uvicorn==0.25.0
fastapi==0.108.0
httpx==0.26.0
